cambio de resources
