+++ .gitignore
@@ -1,0 +1,8 @@
+/bin/
+
+/target/
+
+pom.xml
+.classpath
+.settings/org.eclipse.jdt.core.prefs
+target/.gitignore
