--- src/main/java/tests/TestAdmin.java
@@ -1,5 +1,0 @@
-package tests;
-
-public class TestAdmin {
-
-}
