--- src/main/java/tests/TestCliente.java
@@ -1,5 +1,0 @@
-package tests;
-
-public class TestCliente {
-
-}
