--- src/main/java/tests/TestUsuarios.java
@@ -1,5 +1,0 @@
-package tests;
-
-public class TestUsuarios {
-
-}
