+++ src/test/java/tests/TestAdmin.java
@@ -1,0 +1,5 @@
+package tests;
+
+public class TestAdmin {
+
+}
